+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.19.0] - 2025-02-28
+
+### Added
+- Added support for `bank_transaction_id` parameter in PaymentBase and Transaction::Query classes
+- Added tests for `bank_transaction_id` functionality
+
 ## [0.18.0] - 2025-02-05
 
 ### Added